+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 0.1.7
 - Fix holes in exported terrain being slightly offset.
-- Fix terrain material name conflicts in by using a unique name per tile.
+- Fix terrain material name conflicts in Blender by using a unique name per tile.
 
 0.1.6 (03-01-2020)
 - Fix crash when exporting modern ADTs.