+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 0.1.7
+- Users will now be prompted when a new version of the Blender add-on is available.
+- Automatic installation of the Blender add-on now supports alpha/beta versions of Blender.
 - Fix holes in exported terrain being slightly offset.
 - Fix terrain material name conflicts in Blender by using a unique name per tile.